Renamed all *.js files to *.jsx
